requirements.txt
mysql-connector-python==8.0.33
tkinter
matplotlib
fpdf2